internship_project/requirements.txt
Flask
requests
python-dotenv
transformers
torch
sentencepiece
accelerate
google-generativeai


